django_app/requirements.txt
Django==3.0.5
djangorestframework==3.11.0
gunicorn==20.0.4
redis==3.5.0